Move get-user-details endpoint from /api/auth to /api/user
