+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/config/SecurityConfig.java
@@ -30,7 +30,7 @@ public class SecurityConfig {
                     .requestMatchers("/api/auth/login").permitAll()
                     .requestMatchers("/api/follow/**").permitAll()
                     .requestMatchers("/api/auth/validate-token").permitAll()
-                    .requestMatchers("/api/auth/get-user-details").permitAll()
+                    .requestMatchers("/api/user/get-user-details").permitAll()
                     .anyRequest().authenticated()
             );
         return http.build();
